Update gpcp.xml
Changes based on the meeting earlier today (8-28-2025).

Restored LMS and XMSS to FCS_COP.1/SigGen and FCS_CKM.1/AKG. (Issue 83)

Changed text in FIA_AFL_EXT.1.3 to clarify intent of the requirement for remembering auth failures across reboots. 1.3 and 1.5 may be satisfied by default on devices that boot slow and are configured with lockouts shorter than the boot time. The requirement isn't meant to make an assumption that all devices are slow to boot or will have lockout times shorter than the boot cycle. (Issue 64)